requirements.txt

fastapi
uvicorn
python-jose
sqlalchemy
passlib[argon2]
pyotp
pydantic